job-queue/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
einops==0.8.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sympy==1.13.1
timm==1.0.14
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.3.0
